Don't fail if tac already is symlinked

+++ install/homebrew.sh
@@ -33,7 +33,10 @@ ok "brew upgrade..."
 # Install other common packages
 action "installing packages..."
 require_brew coreutils
-ln -s /usr/local/bin/gtac /usr/local/bin/tac
+if test ! -e /usr/local/bin/tac; then
+    action "linking gtac to tac"
+    ln -s /usr/local/bin/gtac /usr/local/bin/tac
+fi
 require_brew findutils
 #require_brew kdiff3
 require_brew trash
